Refactored to NavHost/Controller
